refactor(context): 复用默认知识库选择规则
